+++ src/main/java/com/smartsheet/api/models/enums/UpgradeSeatType.java
@@ -16,6 +16,6 @@
 
 package com.smartsheet.api.models.enums;
 
-public enum SeatType {
-    MEMBER, PROVISIONAL_MEMBER, GUEST, VIEWER
+public enum UpgradeSeatType {
+    GUEST, MEMBER
 }